model: move TabModelStore to package persistence
This class is more like a helper from persistence package to help other
classes on storing/restoring TabModel.
