+++ app/src/main/java/org/mozilla/focus/persistence/TabModelStore.java
@@ -1,4 +1,4 @@
-package org.mozilla.focus.tabs;
+package org.mozilla.focus.persistence;
 
 import android.content.Context;
 import android.os.AsyncTask;
@@ -8,8 +8,6 @@ import android.support.annotation.Nullable;
 
 import org.mozilla.focus.Inject;
 import org.mozilla.focus.R;
-import org.mozilla.focus.persistence.TabEntity;
-import org.mozilla.focus.persistence.TabsDatabase;
 import org.mozilla.focus.utils.FileUtils;
 
 import java.io.File;
